Add Fabric support and update to Dynmap 3.8 and the latest Chunky 2.5.0 snapshot  (#68)
* Move files to subproject

* Move files to common dir

* Add fabric 1.21.11 support

* Fix mod icon

* Use leMaik's maven repo instead of local files for dynmap and chunky-core.

* Update log adapter to include latest upstream changes.

* Simplify rs-map-icon extraction code.

* Indent all gradle files with spaces instead of tabs.

---------
