+++ .gitignore
@@ -2,3 +2,6 @@
 target
 *.iml
 *.jar
+
+.gradle
+build
