+++ chunkymap-core/src/main/java/de/lemaik/chunkymap/dynmap/DynmapCameraAdapter.java
