+++ chunkymap-core/src/main/resources/rs-map-icon.png
