+++ chunkymap-fabric/gradle.properties
@@ -1,0 +1,8 @@
+# Fabric Properties
+# check these on https://fabricmc.net/develop
+minecraft_version=1.21.11
+loader_version=0.18.4
+loom_version=1.15-SNAPSHOT
+
+# Dependencies
+fabric_api_version=0.141.3+1.21.11
