+++ gradle.properties
@@ -1,0 +1,13 @@
+# Done to increase the memory available to gradle.
+org.gradle.jvmargs=-Xmx1G
+org.gradle.parallel=true
+
+# IntelliJ IDEA is not yet fully compatible with configuration cache, see: https://github.com/FabricMC/fabric-loom/issues/1349
+org.gradle.configuration-cache=false
+
+# Project Properties
+version=2.7.0-pre2
+group=de.lemaik.chunkymap
+archives_base_name=chunkymap
+
+chunky_version=2.5.0-SNAPSHOT.446.ga4b37eb
