--- pom.xml
@@ -1,139 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <groupId>de.lemaik.chunkymap</groupId>
-  <artifactId>ChunkyMap</artifactId>
-  <version>2.7.0-pre1</version>
-
-  <licenses>
-    <license>
-      <name>GPL v3</name>
-      <url>https://www.gnu.org/licenses/gpl-3.0.txt</url>
-    </license>
-  </licenses>
-
-  <properties>
-    <jdk.version>21</jdk.version>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-  </properties>
-
-  <repositories>
-    <repository>
-      <id>sonatype-oss-releases</id>
-      <url>https://oss.sonatype.org/content/repositories/releases</url>
-    </repository>
-    <repository>
-      <id>spigot-repo</id>
-      <url>https://hub.spigotmc.org/nexus/content/groups/public/</url>
-    </repository>
-    <repository>
-      <id>wertarbyte-repo</id>
-      <url>https://repo.wertarbyte.com</url>
-    </repository>
-  </repositories>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.spigotmc</groupId>
-      <artifactId>spigot-api</artifactId>
-      <version>1.19.2-R0.1-SNAPSHOT</version>
-      <scope>provided</scope>
-    </dependency>
-    <dependency>
-      <groupId>se.llbit</groupId>
-      <artifactId>chunky-core</artifactId>
-      <version>2.5.0-DEV.381.gf277ba6</version>
-    </dependency>
-    <dependency>
-      <!-- dependency of chunky -->
-      <groupId>org.apache.commons</groupId>
-      <artifactId>commons-math3</artifactId>
-      <version>3.2</version>
-    </dependency>
-    <dependency>
-      <groupId>us.dynmap</groupId>
-      <artifactId>DynmapCore</artifactId>
-      <version>3.7-beta-8</version>
-      <scope>provided</scope>
-      <exclusions>
-        <exclusion>
-          <!-- downloading fails -->
-          <groupId>javax.servlet</groupId>
-          <artifactId>javax.servlet-api</artifactId>
-        </exclusion>
-      </exclusions>
-    </dependency>
-    <dependency>
-      <groupId>us.dynmap</groupId>
-      <artifactId>DynmapCoreAPI</artifactId>
-      <version>3.7-beta-8</version>
-      <scope>provided</scope>
-    </dependency>
-    <dependency>
-      <groupId>com.squareup.okhttp3</groupId>
-      <artifactId>okhttp</artifactId>
-      <version>3.14.9</version>
-    </dependency>
-    <dependency>
-      <groupId>com.google.code.gson</groupId>
-      <artifactId>gson</artifactId>
-      <version>2.8.9</version>
-    </dependency>
-    <dependency>
-      <groupId>net.time4tea</groupId>
-      <artifactId>oidnjni</artifactId>
-      <version>0.1.13</version>
-    </dependency>
-    <dependency>
-      <!-- dependency of oidnjni -->
-      <groupId>org.jetbrains.kotlin</groupId>
-      <artifactId>kotlin-stdlib</artifactId>
-      <version>1.3.72</version>
-    </dependency>
-    <dependency>
-      <groupId>de.lemaik</groupId>
-      <artifactId>chunky-denoiser</artifactId>
-      <version>0.5.0-pre3</version>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-compiler-plugin</artifactId>
-        <version>3.13.0</version>
-        <configuration>
-          <source>21</source>
-          <target>21</target>
-        </configuration>
-      </plugin>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-assembly-plugin</artifactId>
-        <version>2.6</version>
-        <executions>
-          <execution>
-            <phase>package</phase>
-            <goals>
-              <goal>single</goal>
-            </goals>
-          </execution>
-        </executions>
-        <configuration>
-          <descriptorRefs>
-            <descriptorRef>jar-with-dependencies</descriptorRef>
-          </descriptorRefs>
-        </configuration>
-      </plugin>
-    </plugins>
-    <resources>
-      <resource>
-        <directory>src/main/resources</directory>
-        <filtering>true</filtering>
-      </resource>
-    </resources>
-  </build>
-</project>
