+++ settings.gradle
@@ -1,0 +1,13 @@
+pluginManagement {
+	repositories {
+		maven {
+			name = 'Fabric'
+			url = 'https://maven.fabricmc.net/'
+		}
+		mavenCentral()
+		gradlePluginPortal()
+	}
+}
+
+rootProject.name = "chunkymap"
+include("chunkymap-core", "chunkymap-spigot", "chunkymap-fabric")
