--- src/main/resources/plugin.yml
@@ -1,5 +1,0 @@
-name: ChunkyMap
-version: ${project.version}
-main: de.lemaik.chunkymap.ChunkyMapPlugin
-loadbefore: [dynmap]
-softdepend: [dynmap]
